[.-] [example] No ": " prompt..
Printing prompt when needed is actually harder than it seemed.

1. We are not differentiating between spaces/newlines.
2. arduino-cli monitor sends data line-wise.

We can safely print prompt at start.
Then we will receive chunk of data.
  Each WaitEntity() will eat piece of that chunk.
  We should not print prompt till chunk is eaten.
When chunk is eaten, execution point is inside WaitEntity(),
  too late to print prompt.

We can implement fancy inactivity timeouts but for the sake of
simplicity (and sanity!) I'll leave example as is.

(User-friendliness is forced behavior. Code intention is to get data
from other code on another side of that UART channel.)